Make package OSX compatible

+++ pearl-config/install.sh
@@ -1,14 +1,26 @@
 
 function post_install(){
     info "Installing or updating the powerline git repository..."
-    install_or_update_git_repo https://github.com/Lokaltog/powerline.git "${PEARL_PKGVARDIR}/powerline" master
-    mkdir -p ${HOME}/.fonts
-    mkdir -p ${HOME}/.fonts.conf.d/
-    local fonts_file="${PEARL_PKGVARDIR}/powerline/font/PowerlineSymbols.otf"
-    [ -f "$fonts_file" ] || ln -s "$fonts_file" "${HOME}/.fonts"
-    fc-cache -vf "${HOME}/.fonts"
+    install_or_update_git_repo https://github.com/powerline/powerline.git "${PEARL_PKGVARDIR}/powerline" master
 
-    cp ${PEARL_PKGVARDIR}/powerline/font/10-powerline-symbols.conf ${HOME}/.fonts.conf.d/
+    if command -v fc-cache &> /dev/null
+    then
+        mkdir -p ${HOME}/.fonts
+        mkdir -p ${HOME}/.fonts.conf.d/
+        local fonts_file="${PEARL_PKGVARDIR}/powerline/font/PowerlineSymbols.otf"
+        if [[ -f "$fonts_file" ]]
+        then
+            ln -s "$fonts_file" "${HOME}/.fonts/"
+        else
+            warn "Powerline symbols in powerline source code not found."
+        fi
+        fc-cache -vf "${HOME}/.fonts" || warn "fc-cache command did not work. Skipping the configuration of fonts..."
+        cp ${PEARL_PKGVARDIR}/powerline/font/10-powerline-symbols.conf ${HOME}/.fonts.conf.d/
+    else
+        info "If running in OSX, make sure to follow the procedure for installing fonts:"
+        info "    https://powerline.readthedocs.io/en/latest/installation/osx.html#fonts-installation"
+    fi
+
 
     info "Vim binding applied."
 
@@ -30,10 +42,12 @@ function post_update(){
 }
 
 function pre_remove(){
-    rm -f "${HOME}/.fonts/PowerlineSymbols.otf"
-    fc-cache -vf ${HOME}/.fonts
-
-    rm -f "${HOME}/.fonts.conf.d/10-powerline-symbols.conf"
+    if command -v fc-cache &> /dev/null
+    then
+        rm -f "${HOME}/.fonts/PowerlineSymbols.otf"
+        fc-cache -vf ${HOME}/.fonts
+        rm -f "${HOME}/.fonts.conf.d/10-powerline-symbols.conf"
+    fi
 
     unlink "bash" "${PEARL_PKGVARDIR}/powerline/powerline/bindings/bash/powerline.sh"
     unlink "tmux" "${PEARL_PKGVARDIR}/powerline/powerline/bindings/tmux/powerline.conf"
